copyc_svr/requirements.txt
asgiref==3.5.0
Django==3.0
django-ckeditor==6.2.0
django-filter==21.1
django-js-asset==1.2.2
djangorestframework==3.13.1
drf-generators==0.5.0
importlib-metadata==4.10.1
Markdown==3.3.6
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
zipp==3.7.0
